refactor(app): integrated new location button into select-zone
